+++ ndlib/ProfileModel.py
@@ -35,7 +35,7 @@ class ProfileModel(DiffusionModel):
 
             if infected > 0:
                 eventp = np.random.random_sample()
-                if eventp >= self.params['nodes']['profile'][u]:
+                if eventp >= self.params['nodes']['profile'][str(u)]:
                     actual_status[u] = 1
 
         delta = self.status_delta(actual_status)